fastapi/requirements.txt
fastapi
pillow
python-multipart 
aiofiles
pytorch-lightning
tqdm
matplotlib
scikit-learn
scipy
tifffile
ipywidgets
ipython
uvicorn
starlette